pyproject.toml
[project]
name = "typesense"
description = "Python client for Typesense, an open source and typo tolerant search engine."
authors = [{ name = "Typesense", email = "contact@typesense.org" }]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "Apache 2.0" }
keywords = [
    "search",
    "typesense",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = ["requests","httpx"]
dynamic = ["version"]

[project.urls]
Documentation = "https://typesense.org/"
Source = "https://github.com/typesense/typesense-python"
Tracker = "https://github.com/typesense/typesense-python/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "coverage",
    "pytest-mock",
    "requests-mock",
    "python-dotenv",
    "types-requests",
    "faker",
    "ruff>=0.11.11",
    "isort>=6.0.1",
    "pytest-asyncio",
    "pytest-httpx",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.setuptools.dynamic]
version = {attr = "typesense.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
source = ["."]
omit = ["examples/*.py","./venv/*","tests/*/*.py","*__init__.py","*/*test.py", "./src/typesense/types/*.py"]

[tool.coverage.report]
omit = ["examples/*.py","./venv/*","tests/*.py","*__init__.py","*/*test.py"]
